Evaluation measures simplified.

+++ src/test/java/delfos/results/evaluationmeasures/ratingprediction/NRMSETest.java
@@ -177,8 +177,8 @@ public class NRMSETest {
         MeasureResult result = instance.getMeasureResult(recommendationResults, testDataset, relevanceCriteria);
 
         //Phase 3: Result checking
-        float expResult = 0;
-        float delta = 0.001f;
+        double expResult = 0;
+        double delta = 0.001f;
         assertEquals(expResult, result.getValue(), delta);
     }
 
@@ -206,9 +206,9 @@ public class NRMSETest {
         MeasureResult nRMSEResult = instance.getMeasureResult(recommendationResults, testDataset, relevanceCriteria);
 
         //Phase 3: Result checking
-        float expResult = 0.4330127f;
-        float result = nRMSEResult.getValue();
-        float delta = 0.001f;
+        double expResult = 0.4330127;
+        double result = nRMSEResult.getValue();
+        double delta = 0.001f;
         assertEquals(expResult, result, delta);
     }
 }
